(PC-29656)[PRO] refactor: Remove GetAdressDataAdapter adapter.
