--- pro/src/components/Address/adapter/index.ts
@@ -1,1 +1,0 @@
-export { default as getAdressDataAdapter } from './getAdressDataAdapter'
